requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.1
websockets==13.0.1
gradio_client==0.2.0
